move entity_model functions into IocFactory class
