+++ src/ibek/ioc_cmds/commands.py
@@ -5,13 +5,13 @@ from typing import Annotated, List, Optional
 
 import typer
 
-from ibek.entity_model import ioc_create_model
 from ibek.globals import (
     GLOBALS,
     SUPPORT_YAML_PATTERN,
     NaturalOrderGroup,
 )
 from ibek.ioc_cmds.docker import build_dockerfile
+from ibek.ioc_factory import IocFactory
 
 from .assets import extract_assets
 
@@ -78,7 +78,9 @@ def generate_schema(
         log.error(f"No `definitions` given and none found in {GLOBALS.IBEK_DEFS}")
         raise typer.Exit(1)
 
-    ioc_model, _ = ioc_create_model(definitions)
+    ioc_factory = IocFactory()
+    ioc_model = ioc_factory.ioc_create_model(definitions)
+
     schema = json.dumps(ioc_model.model_json_schema(), indent=2)
     if output is None:
         typer.echo(schema)
