+++ tests/test_unit.py
@@ -4,12 +4,11 @@ Some unit tests for ibek.
 
 from ibek.args import IdArg, ObjectArg
 from ibek.commands import semver_compare
-from ibek.entity_model import make_entity_models, make_ioc_model
 from ibek.ioc import clear_entity_model_ids, id_to_entity
 from ibek.support import Definition, Support
 
 
-def test_object_references():
+def test_object_references(factory):
     """
     Verify the object references are correctly resolved
     """
@@ -31,8 +30,8 @@ def test_object_references():
         ],
     )
 
-    entities = make_entity_models(support)
-    ioc_model = make_ioc_model(entities)
+    entities = factory._make_entity_models(support)
+    ioc_model = factory._make_ioc_model(entities)
     assert entities[0].__definition__ == support.defs[0]
     assert entities[1].__definition__ == support.defs[1]
